requirements.txt
pandas>=2.2
pyarrow>=15.0
requests>=2.31
python-dotenv>=1.0
jinja2>=3.1
pytz>=2024.1
